Monthly stats

+++ db/migrate/20220503151311_create_trips.rb
@@ -5,9 +5,9 @@ class CreateTrips < ActiveRecord::Migration[7.0]
     create_table :trips do |t|
       t.string :start_address, null: false
       t.string :end_address, null: false
-      t.float :price, null: false
+      t.decimal :price, null: false
       t.date :delivery_date, null: false
-      t.float :distance, null: false
+      t.decimal :distance, null: false
       t.references :user, null: false, foreign_key: true
 
       t.timestamps
